Feat modulo professor

+++ AlfabetizaAi/src/main/java/br/com/dbc/vemser/alfabetizai/repository/ModuloRepository.java
@@ -1,27 +1,23 @@
 package br.com.dbc.vemser.alfabetizai.repository;
-
-import br.com.dbc.vemser.alfabetizai.dto.ProfessorDTO;
 import br.com.dbc.vemser.alfabetizai.enums.ClassificacaoModulo;
 import br.com.dbc.vemser.alfabetizai.exceptions.BancoDeDadosException;
+import br.com.dbc.vemser.alfabetizai.exceptions.RegraDeNegocioException;
 import br.com.dbc.vemser.alfabetizai.models.Admin;
+import lombok.extern.slf4j.Slf4j;
 import br.com.dbc.vemser.alfabetizai.models.Modulo;
 import br.com.dbc.vemser.alfabetizai.models.Professor;
-import br.com.dbc.vemser.alfabetizai.services.AdminService;
 import br.com.dbc.vemser.alfabetizai.services.ProfessorService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Repository;
-
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
-
+@Slf4j
 @AllArgsConstructor
 @Repository
 public class ModuloRepository implements Repositorio<Integer, Modulo>{
 
     private final ProfessorService professorService;
-
-    private final ConexaoBancoDeDados conexaoBancoDeDados;
 
     @Override
     public Integer getProximoId(Connection connection) throws SQLException {
@@ -29,24 +25,20 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
 
         Statement stmt = connection.createStatement();
         ResultSet res = stmt.executeQuery(sql);
-
         if (res.next()) {
             return res.getInt("mysequence");
         }
-
         return null;
     }
-
     @Override
     public Integer getProximoIdUsuario(Connection connection) throws BancoDeDadosException {
             return null;
     }
-
     @Override
     public Modulo adicionar(Modulo modulo) throws BancoDeDadosException {
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
 
             Integer proximoId = this.getProximoId(con);
             modulo.setId(proximoId);
@@ -58,14 +50,13 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
             PreparedStatement stmt = con.prepareStatement(sql);
 
             stmt.setInt(1, modulo.getId());
-            stmt.setInt(2, modulo.getAutor().getIdProfessor());
+            stmt.setInt(2, modulo.getIdProfessor());
             stmt.setString(3, modulo.getTitulo());
             stmt.setString(4, modulo.getConteudo());
             stmt.setInt(5, modulo.getClassificacao().ordinal()+1);
             stmt.setString(6, "N");
 
             int res = stmt.executeUpdate();
-            System.out.println("adicionarModulo.res=" + res);
             return modulo;
         } catch (SQLException e) {
             throw new BancoDeDadosException(e.getCause());
@@ -84,21 +75,24 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
     public boolean remover(Integer id) throws BancoDeDadosException {
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
+            if (isModuloVinculado(id)) {
+                throw new RegraDeNegocioException("Módulo vinculado a outros registros. Não pode ser excluído.");
+            }
 
             String sql = "DELETE FROM MODULO WHERE id_modulo = ?";
 
             PreparedStatement stmt = con.prepareStatement(sql);
-
             stmt.setInt(1, id);
 
-            // Executa-se a consulta
             int res = stmt.executeUpdate();
-            System.out.println("removerModuloPorId.res=" + res);
-
-            return res > 0;
+            if (res == 0) {
+                throw new RegraDeNegocioException("Dados do Modulo Não Encontrado para o  ID: " + id);
+            }return res > 0;
         } catch (SQLException e) {
             throw new BancoDeDadosException(e.getCause());
+        } catch (RegraDeNegocioException e) {
+            throw new RuntimeException(e);
         } finally {
             try {
                 if (con != null) {
@@ -109,35 +103,34 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
             }
         }
     }
-
     @Override
     public Modulo editar(Integer id, Modulo modulo) throws BancoDeDadosException {
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
+            String sql = "UPDATE MODULO SET " +
+                    "id_professor = ?, " +
+                    "titulo = ?, " +
+                    "conteudo = ?, " +
+                    "classificacao_modulo = ? " +
+                    "WHERE id_modulo = ?";
 
-            StringBuilder sql = new StringBuilder();
-            sql.append("UPDATE MODULO SET ");
-            sql.append(" id_professor = ?,");
-            sql.append(" titulo = ?,");
-            sql.append(" conteudo = ?,");
-            sql.append(" classificacao_modulo = ? ");
-            sql.append(" WHERE id_modulo = ? ");
-
-            PreparedStatement stmt = con.prepareStatement(sql.toString());
-
-            stmt.setInt(1, modulo.getAutor().getIdProfessor());
+            PreparedStatement stmt = con.prepareStatement(sql);
+            stmt.setInt(1, modulo.getIdProfessor());
             stmt.setString(2, modulo.getTitulo());
             stmt.setString(3, modulo.getConteudo());
-            stmt.setInt(4, modulo.getClassificacao().ordinal()+1);
+            stmt.setInt(4, modulo.getClassificacao().ordinal() + 1);
             stmt.setInt(5, id);
 
-            int res = stmt.executeUpdate();
-            System.out.println("editarModulo.res=" + res);
-
+            if (stmt.executeUpdate() == 0) {
+                throw new RegraDeNegocioException("Dados do Módulo Não Encontrados. ID: " + id);
+            }
+            modulo.setId(id);
             return modulo;
         } catch (SQLException e) {
             throw new BancoDeDadosException(e.getCause());
+        } catch (RegraDeNegocioException e) {
+            throw new RuntimeException(e);
         } finally {
             try {
                 if (con != null) {
@@ -151,7 +144,7 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
     public boolean editarAprovacaoPorAdmin(Integer idAdmin, Integer idModulo, String aprovacaoModulo) throws BancoDeDadosException {
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
 
             StringBuilder sql = new StringBuilder();
             sql.append("UPDATE MODULO SET ");
@@ -160,17 +153,19 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
             sql.append(" WHERE id_modulo = ? ");
 
             PreparedStatement stmt = con.prepareStatement(sql.toString());
-
             stmt.setInt(1, idAdmin);
             stmt.setString(2, aprovacaoModulo);
             stmt.setInt(3, idModulo);
 
             int res = stmt.executeUpdate();
-            System.out.println("editarModulo.res=" + res);
+            if (res == 0) {
+                throw new RegraDeNegocioException("Dados do Admin Não Encontrado. ID: " + idAdmin);
+            }return res > 0;
 
-            return res > 0;
         } catch (SQLException e) {
             throw new BancoDeDadosException(e.getCause());
+        } catch (RegraDeNegocioException e) {
+            throw new RuntimeException(e);
         } finally {
             try {
                 if (con != null) {
@@ -181,44 +176,41 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
             }
         }
     }
-
     @Override
     public List<Modulo> listar() throws Exception {
         List<Modulo> modulos = new ArrayList<>();
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
             Statement stmt = con.createStatement();
-
             String sql = "SELECT * FROM MODULO";
 
             ResultSet res = stmt.executeQuery(sql);
-
             while (res.next()) {
                 Modulo modulo = new Modulo();
                 Professor professor = new Professor();
                 Admin admin = new Admin();
 
                 modulo.setId(res.getInt("id_modulo"));
-                professor.setIdProfessor(res.getInt("id_professor"));
-                ProfessorDTO professorDTO = professorService.buscarProfessorPorId(professor.getIdProfessor());
-                professor.setIdProfessor(professorDTO.getIdProfessor());
-                professor.setNome(professorDTO.getNome());
-                professor.setSobrenome(professorDTO.getSobrenome());
-                professor.setEmail(professorDTO.getEmail());
-                professor.setDescricao(professorDTO.getDescricao());
-                professor.setTelefone(professorDTO.getTelefone());
+                modulo.setIdProfessor(res.getInt("id_professor"));
+                professor.setNome(professor.getNome());
+                professor.setSobrenome(professor.getSobrenome());
+                professor.setEmail(professor.getEmail());
+                professor.setDescricao(professor.getDescricao());
+                professor.setTelefone(professor.getTelefone());
+
                 modulo.setAdminAprova(admin);
                 admin.setIdUsuario(res.getInt("id_admin"));
                 modulo.setTitulo(res.getString("titulo"));
                 modulo.setConteudo(res.getString("conteudo"));
                 modulo.setClassificacao(ClassificacaoModulo.trazEnumPeloOrdinal(res.getInt("classificacao_modulo")));
-                modulo.setFoiAprovado((res.getString("modulo_aprovado").charAt(0)));
+                String moduloAprovadoStr = res.getString("modulo_aprovado");
+                modulo.setFoiAprovado((moduloAprovadoStr != null && moduloAprovadoStr.length() > 0) ? moduloAprovadoStr.charAt(0) : ' ');
                 modulo.setAutor(professor);
                 modulos.add(modulo);
             }
-
         } catch (SQLException e) {
+            log.info("Erro ao executar consulta SQL", e);
             throw new BancoDeDadosException(e.getCause());
         } finally {
             try {
@@ -232,15 +224,61 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
         return modulos;
     }
 
+    public List<Modulo> listarPorIdProfessor(Integer idProfessor) throws BancoDeDadosException {
+        List<Modulo> moduloPorProfessor = new ArrayList<>();
+        Connection con = null;
+        try {
+            con = ConexaoBancoDeDados.getConnection();
+            if (!existeProfessor(idProfessor, con)) {
+                throw new RegraDeNegocioException("Professor com ID " + idProfessor + " não encontrado");
+            }
+            String sql = "SELECT M.ID_MODULO, P.ID_PROFESSOR, M.TITULO, M.CONTEUDO, M.CLASSIFICACAO_MODULO, M.MODULO_APROVADO " +
+                    "FROM MODULO M " +
+                    "JOIN PROFESSOR P ON " +
+                    "M.ID_PROFESSOR = P.ID_PROFESSOR " +
+                    "WHERE P.ID_PROFESSOR = ?";
+
+            PreparedStatement stmt = con.prepareStatement(sql);
+            stmt.setInt(1, idProfessor);
+            ResultSet res = stmt.executeQuery();
+            while (res.next()) {
+                Modulo moduloResponse = new Modulo();
+                moduloResponse.setId(res.getInt("ID_MODULO"));
+                moduloResponse.setIdProfessor(res.getInt("ID_PROFESSOR"));
+                moduloResponse.setTitulo(res.getString("TITULO"));
+                moduloResponse.setConteudo(res.getString("CONTEUDO"));
+                moduloResponse.setClassificacao(ClassificacaoModulo.trazEnumPeloOrdinal(res.getInt(5)));
+                String moduloAprovadoStr = res.getString("modulo_aprovado");
+                moduloResponse.setFoiAprovado((moduloAprovadoStr != null && moduloAprovadoStr.length() > 0) ? moduloAprovadoStr.charAt(0) : ' ');
+                moduloPorProfessor.add(moduloResponse);
+            }
+            if (moduloPorProfessor.isEmpty()) {
+                throw new RegraDeNegocioException("Nenhum modulo " +
+                        "encontrado para o professor com ID: " + idProfessor);
+            }
+            return moduloPorProfessor;
+        } catch (SQLException e) {
+            throw new BancoDeDadosException(e.getCause());
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (con != null) {
+                    con.close();
+                }
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+    }
     public List<Modulo> listarSemAprovacao() throws BancoDeDadosException {
         List<Modulo> modulos = new ArrayList<>();
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
             Statement stmt = con.createStatement();
 
             String sql = "SELECT * FROM MODULO WHERE MODULO_APROVADO = 'N' OR MODULO_APROVADO = 'n'";
-
             ResultSet res = stmt.executeQuery(sql);
 
             while (res.next()) {
@@ -266,16 +304,14 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
         }
         return modulos;
     }
-
     public List<Modulo> listarAprovados() throws BancoDeDadosException {
         List<Modulo> modulos = new ArrayList<>();
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
             Statement stmt = con.createStatement();
 
             String sql = "SELECT * FROM MODULO WHERE MODULO_APROVADO = 'S' OR MODULO_APROVADO = 's'";
-
             ResultSet res = stmt.executeQuery(sql);
 
             while (res.next()) {
@@ -307,16 +343,14 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
         }
         return modulos;
     }
-
     public List<Modulo> listarReprovados() throws BancoDeDadosException {
         List<Modulo> modulos = new ArrayList<>();
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
             Statement stmt = con.createStatement();
 
             String sql = "SELECT * FROM MODULO WHERE MODULO_APROVADO = 'R' OR MODULO_APROVADO = 'r'";
-
             ResultSet res = stmt.executeQuery(sql);
 
             while (res.next()) {
@@ -331,7 +365,8 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
                 modulo.setTitulo(res.getString("titulo"));
                 modulo.setConteudo(res.getString("conteudo"));
                 modulo.setClassificacao(ClassificacaoModulo.trazEnumPeloOrdinal(res.getInt("classificacao_modulo")));
-                modulo.setFoiAprovado((res.getString("modulo_aprovado").charAt(0)));
+                String moduloAprovadoStr = res.getString("modulo_aprovado");
+                modulo.setFoiAprovado((moduloAprovadoStr != null && moduloAprovadoStr.length() > 0) ? moduloAprovadoStr.charAt(0) : ' ');
                 modulos.add(modulo);
             }
 
@@ -351,7 +386,7 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
     public Modulo buscarModuloPorId(Integer idModulo) throws Exception {
         Connection con = null;
         try {
-            con = conexaoBancoDeDados.getConnection();
+            con = ConexaoBancoDeDados.getConnection();
 
             String sql = "SELECT * FROM MODULO M " +
                     "WHERE M.ID_MODULO = ?";
@@ -360,27 +395,20 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
             stmt.setInt(1, idModulo);
 
             ResultSet res = stmt.executeQuery();
-            Modulo modulo = new Modulo();
-            while (res.next()) {
-                ProfessorDTO professorDTO = professorService.buscarProfessorPorId(res.getInt("id_professor"));
-                Professor professor = new Professor();
-                professor.setIdProfessor(professorDTO.getIdProfessor());
-                professor.setNome(professorDTO.getNome());
-                professor.setSobrenome(professorDTO.getSobrenome());
-                professor.setTelefone(professorDTO.getTelefone());
-                professor.setEmail(professorDTO.getEmail());
-                professor.setDescricao(professorDTO.getDescricao());
-                modulo.setId(res.getInt("id_modulo"));
-                modulo.setAutor(professor);
-                modulo.setTitulo(res.getString("titulo"));
-                modulo.setConteudo(res.getString("conteudo"));
-                modulo.setClassificacao(ClassificacaoModulo.ofTipo(res.getInt("classificacao_modulo")));
-                modulo.setFoiAprovado(res.getString("modulo_aprovado").charAt(0));
-
+            if (!res.next()) {
+                throw new RegraDeNegocioException("Dados do Módulo Não Encontrado para o ID: " + idModulo);
             }
-            System.out.println(modulo);
-            return modulo;
+            Modulo modulo = new Modulo();
 
+            modulo.setId(res.getInt("id_modulo"));
+            modulo.setIdProfessor(res.getInt("id_professor"));
+            modulo.setTitulo(res.getString("titulo"));
+            modulo.setConteudo(res.getString("conteudo"));
+            modulo.setClassificacao(ClassificacaoModulo.ofTipo(res.getInt("classificacao_modulo")));
+            String moduloAprovadoStr = res.getString("modulo_aprovado");
+            modulo.setFoiAprovado((moduloAprovadoStr != null && moduloAprovadoStr.length() > 0) ? moduloAprovadoStr.charAt(0) : ' ');
+
+            return modulo;
         } catch (SQLException e) {
             e.printStackTrace();
             throw new BancoDeDadosException(e.getCause());
@@ -392,6 +420,35 @@ public class ModuloRepository implements Repositorio<Integer, Modulo>{
                 }
             } catch (SQLException e) {
                 e.printStackTrace();
+            }
+        }
+    }
+    private boolean existeProfessor(Integer idProfessor, Connection connection) throws SQLException {
+        String sql = "SELECT * FROM PROFESSOR WHERE ID_PROFESSOR  = ?";
+        try (PreparedStatement stmt = connection.prepareStatement(sql)) {
+            stmt.setInt(1, idProfessor);
+            ResultSet resultSet = stmt.executeQuery();
+            return resultSet.next();
+        }
+}
+    private boolean isModuloVinculado(Integer id) throws SQLException {
+        Connection con = null;
+        try {
+            con = ConexaoBancoDeDados.getConnection();
+
+            String sql = "SELECT COUNT(*) FROM DESAFIO WHERE id_modulo = ?";
+            PreparedStatement stmt = con.prepareStatement(sql);
+            stmt.setInt(1, id);
+
+            ResultSet resultSet = stmt.executeQuery();
+            if (resultSet.next()) {
+                int count = resultSet.getInt(1);
+                return count > 0;
+            }
+            return false;
+        } finally {
+            if (con != null) {
+                con.close();
             }
         }
     }
